Move flattenFloats function to a separate file
