+++ api/src/test/kotlin/org/jetbrains/kotlinx/dl/api/core/layer/LayerTest.kt
@@ -5,9 +5,9 @@
 
 package org.jetbrains.kotlinx.dl.api.core.layer
 
-import org.jetbrains.kotlinx.dl.api.core.shape.flattenFloats
 import org.jetbrains.kotlinx.dl.api.core.shape.shape
 import org.jetbrains.kotlinx.dl.api.core.shape.toLongArray
+import org.jetbrains.kotlinx.dl.api.core.util.flattenFloats
 import org.jetbrains.kotlinx.dl.api.extension.convertTensorToFlattenFloatArray
 import org.junit.jupiter.api.Assertions.assertArrayEquals
 import org.tensorflow.*
